DBProvider 업그레이드 로직 개선

+++ JustViewer2/app/src/main/java/com/kds3393/just/justviewer2/db/DBProvider.kt
@@ -34,11 +34,18 @@ class DBProvider(context: Context?, path: String?) {
         }
 
         override fun onUpgrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
-            db.execSQL("DROP TABLE IF EXISTS " + DBInfo.FAVO_BOOKMARK_TABLE)
-            db.execSQL("DROP TABLE IF EXISTS image_bookmark_table") // 구형 테이블
-            db.execSQL("DROP TABLE IF EXISTS txt_bookmark_table")   // 구형 테이블
-            db.execSQL("DROP TABLE IF EXISTS " + DBInfo.BOOKMARK_TABLE)
-            onCreate(db)
+            if (oldVersion < newVersion) {
+                try {
+                    db.execSQL("DROP TABLE IF EXISTS ${DBInfo.FAVO_BOOKMARK_TABLE}")
+                    db.execSQL("DROP TABLE IF EXISTS ${DBInfo.BOOKMARK_TABLE}")
+                    db.execSQL("DROP TABLE IF EXISTS image_bookmark_table")
+                    db.execSQL("DROP TABLE IF EXISTS txt_bookmark_table")
+
+                    onCreate(db)
+                } catch (e: SQLiteException) {
+                    CLog.e(e)
+                }
+            }
         }
     }
